requirements.txt
Flask
python-whois
gunicorn
shodan